test: move test class to test/java/domain package
