+++ src/test/java/domain/LottoNumberStrategyTest.java
@@ -1,10 +1,9 @@
-import domain.RandomLottoNumberStrategy;
-import domain.LottoNumberStrategy;
+package domain;
+
+import static org.assertj.core.api.AssertionsForClassTypes.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 class LottoNumberStrategyTest {
