+++ src/test/java/domain/LottoRankTest.java
@@ -1,11 +1,11 @@
+package domain;
+
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-
-import domain.LottoRank;
 
 class LottoRankTest {
