+++ src/test/java/domain/MoneyTest.java
@@ -1,9 +1,10 @@
-import domain.Money;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
+package domain;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 
 class MoneyTest {
